requirements.txt
amqp==1.4.9
anyjson==0.3.3
astroid==2.0.4
autopep8==1.4.4
billiard==3.3.0.23
cairocffi==1.0.2
CairoSVG==2.4.0
celery==3.1.24
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
colorama==0.3.9
crypto==1.4.1
cssselect2==0.2.1
decorator==4.3.0
defusedxml==0.6.0
Django==2.1.7
django-ckeditor==5.6.1
django-cors-headers==2.5.2
django-crispy-forms==1.7.2
django-haystack==2.8.1
django-js-asset==1.1.0
django-materialize==1.0.1a2
django-taggit==0.23.0
djangorestframework==3.8.2
forex-python==1.1
html5lib==1.0.1
idna==2.8
isort==4.3.4
jsonpickle==1.0
kombu==3.0.37
lazy-object-proxy==1.3.1
mccabe==0.6.1
Naked==0.1.31
paystackapi==1.2.7
paywix==2.1
Pillow==5.2.0
psycopg2==2.7.5
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.8.2
pycryptodomex==3.8.2
pylint==2.1.1
pypaystack==1.18
Pyphen==0.9.5
pysolr==3.8.1
python-paystack==1.2.0
pytz==2018.5
PyYAML==3.13
rave-python==1.0.1
redis==3.2.1
reportlab==3.5.23
requests==2.21.0
shellescape==3.4.1
simplejson==3.16.0
six==1.11.0
sorl-thumbnail==12.5.0
tinycss2==1.0.2
urllib3==1.24.1
validators==0.12.4
vine==1.1.4
WeasyPrint==48
webencodings==0.5.1
wrapt==1.10.11
